+++ src/Extension.php
@@ -257,12 +257,36 @@ class Pronamic_WP_Pay_Extensions_GravityForms_Extension {
 			// @see https://github.com/gravityforms/gravityformsaweber/blob/1.4.2/aweber.php#L1167-L1197
 			if ( $feed->delay_aweber_subscription && Pronamic_WP_Pay_Class::method_exists( 'GFAWeber', 'export' ) ) {
 				call_user_func( array( 'GFAWeber', 'export' ), $entry, $form, false );
+
+				// @since 1.3.0
+				// @see https://github.com/gravityforms/gravityformsaweber/blob/2.2.1/aweber.php#L48-L50
+				// @see https://github.com/gravityforms/gravityforms/blob/1.9.10.15/includes/addon/class-gf-feed-addon.php#L43
+				if ( function_exists( 'gf_aweber' ) ) {
+					$addon = gf_aweber();
+
+					if ( method_exists( $addon, 'maybe_process_feed' ) ) {
+						$addon->maybe_process_feed( $entry, $form );
+					}
+				}
 			}
 
 			// Delay Campaign Monitor
-			// @see https://github.com/gravityforms/gravityformscampaignmonitor/blob/2.5.1/campaignmonitor.php#L1184
-			if ( $feed->delay_campaignmonitor_subscription && Pronamic_WP_Pay_Class::method_exists( 'GFCampaignMonitor', 'export' ) ) {
-				call_user_func( array( 'GFCampaignMonitor', 'export' ), $entry, $form, false );
+			if ( $feed->delay_campaignmonitor_subscription ) {
+				// @see https://github.com/gravityforms/gravityformscampaignmonitor/blob/2.5.1/campaignmonitor.php#L1184
+				if ( Pronamic_WP_Pay_Class::method_exists( 'GFCampaignMonitor', 'export' ) ) {
+					call_user_func( array( 'GFCampaignMonitor', 'export' ), $entry, $form, false );
+				}
+
+				// @since 1.3.0
+				// @see https://github.com/gravityforms/gravityformscampaignmonitor/blob/3.3.2/campaignmonitor.php#L48-L50
+				// @see https://github.com/gravityforms/gravityforms/blob/1.9.10.15/includes/addon/class-gf-feed-addon.php#L43
+				if ( function_exists( 'gf_campaignmonitor' ) ) {
+					$addon = gf_campaignmonitor();
+
+					if ( method_exists( $addon, 'maybe_process_feed' ) ) {
+						$addon->maybe_process_feed( $entry, $form );
+					}
+				}
 			}
 
 			// Delay Mailchimp
@@ -273,9 +297,10 @@ class Pronamic_WP_Pay_Extensions_GravityForms_Extension {
 				}
 
 				// @since 1.3.0
-				// @see https://github.com/gravityforms/gravityforms/blob/1.9.10.15/includes/addon/class-gf-feed-addon.php#L140-L225
-				if ( Pronamic_WP_Pay_Class::method_exists( 'GFMailChimp', 'get_instance' ) ) {
-					$addon = GFMailChimp::get_instance();
+				// @see https://github.com/gravityforms/gravityformsmailchimp/blob/3.6.3/mailchimp.php#L48-L50
+				// @see https://github.com/gravityforms/gravityforms/blob/1.9.10.15/includes/addon/class-gf-feed-addon.php#L43
+				if ( function_exists( 'gf_mailchimp' ) ) {
+					$addon = gf_mailchimp();
 
 					if ( method_exists( $addon, 'maybe_process_feed' ) ) {
 						$addon->maybe_process_feed( $entry, $form );